PythonModule/controller/requirements.txt
numpy
pandas
matplotlib
pyCandleMAB==1.5.0

